fix: apply code review changes
